% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{customerExp}
\alias{customerExp}
\title{Customer's job size and patience
Provides with a realization of Y~Exp(theta) and B~Gamma(eta,mu)}
\usage{
customerExp(params)
}
\arguments{
\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
list with elements 'Patience' and 'Jobsize'
}
\description{
Customer's job size and patience
Provides with a realization of Y~Exp(theta) and B~Gamma(eta,mu)
}
\examples{
params <- listParams(gamma=10,lambda_0=20,theta=2.5, eta = 1, mu = 1 , s = 3)
 customer <- customerExp(params)
 customer
}
