% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleDataAWX}
\alias{exampleDataAWX}
\title{Auxiliary function to create example dataset}
\usage{
exampleDataAWX()
}
\value{
AWX dataset with n = 1000 generated with exampleParams()
}
\description{
Auxiliary function to create example dataset
}
\examples{
set.seed(123)
dat <- exampleDataAWX()
set.seed(123)
pars <- exampleParams()
dat2 <- resSimAWX(n_thousands = 1,params = pars)
identical(dat,dat2)
}
