% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleDataRES}
\alias{exampleDataRES}
\title{Auxiliary function to create example full data}
\usage{
exampleDataRES(scenario = "C4")
}
\arguments{
\item{scenario}{string with scenario names}
}
\value{
RES list with n = 1000 generated with exampleParams()
}
\description{
Auxiliary function to create example full data
}
