% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{exampleParams}
\alias{exampleParams}
\title{Auxiliary function to create example parameters}
\usage{
exampleParams()
}
\value{
a list of example parameters
}
\description{
Auxiliary function to create example parameters
}
\examples{
params <- listParams(gamma=1,lambda_0=2,theta=1, eta = 1, mu = 1 , s = 3)
identical(params,exampleParams())
}
