% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{filenamer}
\alias{filenamer}
\title{Utility: name the individual simulation results}
\usage{
filenamer(n_obs, params)
}
\arguments{
\item{n_obs}{number of observations in data}

\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
a time-stamped filename for a simulation dataframe
}
\description{
Utility: name the individual simulation results
}
\examples{
filenamer(1000,params = exampleParams())
}
