% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GL0T}
\alias{GL0T}
\title{get gamma, lambda_0 and theta}
\usage{
GL0T(params)
}
\arguments{
\item{params}{names list of parameters (output of 'listParams()').}
}
\value{
a named vector (gamma,lambda_0,theta)
}
\description{
get gamma, lambda_0 and theta
}
\examples{
GL0T(exampleParams())
}
