% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{grad.cheap}
\alias{grad.cheap}
\title{Cheap evaluation of the gradient}
\usage{
grad.cheap(params, which_known, AWX)
}
\arguments{
\item{params}{list of parameters}

\item{which_known}{vector of length 1-3, 1 = d_gamma 2 = d_lambda_0 3 = d_theta}

\item{AWX}{data}
}
\value{
a named vector with the corresponding gradient for the unknown parameters
}
\description{
Cheap evaluation of the gradient
}
