% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{grad}
\alias{grad}
\title{generic gradient function}
\usage{
grad(point, params, AWX)
}
\arguments{
\item{point}{named vector of 2-3 of parameters (gamma,lambda_0,theta)}

\item{params}{the actual parameter values}

\item{AWX}{the dataset}
}
\value{
a named vector with the value of the gradient at point, assuming the
unspecified parameter is \emph{known!}
}
\description{
generic gradient function
}
\examples{

params <- exampleParams()
names(GL0T(params))
AWX <- exampleDataAWX()
point <- GL0T(params)[c("lambda_0","theta")]
grad(point = point,AWX = AWX,params = params)
point <- GL0T(params)[c("lambda_0","gamma")]
grad(point = point,AWX = AWX,params = params)
point <- GL0T(params)[c("gamma","lambda_0")]
grad(point = point,AWX = AWX,params = params)
point <- GL0T(params) + runif(3)
grad(point = point,AWX = AWX,params = params)
}
