% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gradLogLik}
\alias{gradLogLik}
\title{Gradient of the full likelihood, or a partial version}
\usage{
gradLogLik(gamma, lambda_0, theta, AWX)
}
\arguments{
\item{gamma}{gamma}

\item{lambda_0}{lambda_0}

\item{theta}{theta}

\item{AWX}{AWX}
}
\value{
the gradient at (gamma,lambda_0,theta) for data AWX
}
\description{
Gradient of the full likelihood, or a partial version
}
