% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gradLogLik.vec}
\alias{gradLogLik.vec}
\title{Gradient of  likelihood, vector form}
\usage{
gradLogLik.vec(g_l0_t, AWX)
}
\arguments{
\item{g_l0_t}{vector c(gamma, lambda_0, theta)}

\item{AWX}{AWX}
}
\value{
the gradient at (gamma,lambda_0,theta) for data AWX
}
\description{
Gradient of  likelihood, vector form
}
