% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gradOneKnown}
\alias{gradOneKnown}
\title{gradient when one parameter is known}
\usage{
gradOneKnown(point, params, AWX)
}
\arguments{
\item{point}{named vector of the \emph{uknown} parameters.}

\item{params}{simulation parameters}

\item{AWX}{dataset}
}
\value{
The gradient at the supplied point
}
\description{
gradient when one parameter is known
}
