% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{makeAWXDirectories}
\alias{makeAWXDirectories}
\title{Interactive function that creates an entire scenario}
\usage{
makeAWXDirectories()
}
\value{


Returns NULL. Creates folders with the realizations for each value of s.
}
\description{
Interactive function that creates an entire scenario
}
\details{
User is prompted for the parameters in the console.
}
