% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleAll}
\alias{mleAll}
\title{Compute all MLE's for a dataset}
\usage{
mleAll(AWX, params)
}
\arguments{
\item{AWX}{can be a dataframe or a path}

\item{params}{parameter list}
}
\value{
named vector of all MLE's
}
\description{
Compute all MLE's for a dataset
}
