% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleFull}
\alias{mleFull}
\title{Maximum likelihood estimation}
\usage{
mleFull(AWX, params)
}
\arguments{
\item{AWX}{AWX dataframe}

\item{params}{vector c(gamma, lambda_0, theta)}
}
\value{
A list with elements: ans - the mle's, boundary = logical indicating whether solution is on boundary
}
\description{
Maximum likelihood estimation
}
