% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleGetAll}
\alias{mleGetAll}
\title{get all MLE's for dataset}
\usage{
mleGetAll(AWX, params)
}
\arguments{
\item{AWX}{dataframe}

\item{params}{parameter list}
}
\value{
a named vector with all possible MLE's
}
\description{
get all MLE's for dataset
}
