% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleLiron}
\alias{mleLiron}
\title{Liron MLE}
\usage{
mleLiron(AWX, acc = 1e-04)
}
\arguments{
\item{AWX}{data}

\item{acc}{accuracy for binary search}
}
\value{
A vector of theta and lambda estimates
}
\description{
Liron MLE
}
