% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleOneKnown}
\alias{mleOneKnown}
\title{Mle with one known parameter}
\usage{
mleOneKnown(AWX, params, which_known)
}
\arguments{
\item{AWX}{data}

\item{params}{parameters}

\item{which_known}{name of the \emph{known} parameter}
}
\value{
Maximum likelihood estimates when one known parameter
}
\description{
Mle with one known parameter
}
