% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{mleTwoKnown}
\alias{mleTwoKnown}
\title{Mle with two known parameters}
\usage{
mleTwoKnown(AWX, params, which_not_known)
}
\arguments{
\item{AWX}{data}

\item{params}{parameters}

\item{which_not_known}{name of the UNKNOWN parameter to be estimated}
}
\value{
Maximum likelihood estimates when one known parameter
}
\description{
Mle with two known parameters
}
