% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{negLogLik}
\alias{negLogLik}
\title{(negative) mean log-likelihood for the sinusoidal arrivals model
The mean is returned instead of the sum - should help gradient based optimizers}
\usage{
negLogLik(gamma, lambda_0, theta, AWX)
}
\arguments{
\item{gamma}{periodical component of rate function}

\item{lambda_0}{constant component of rate function}

\item{theta}{parameter of exponential patience}

\item{AWX}{compact simulation results for memory economy (A,W,X).}
}
\value{
The negative log-likelihood at the point provided. Based on mean (instead of sum).
}
\description{
(negative) mean log-likelihood for the sinusoidal arrivals model
The mean is returned instead of the sum - should help gradient based optimizers
}
