% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{negLogLik.vec}
\alias{negLogLik.vec}
\title{Wrapper for negLogLik to use with optim}
\usage{
negLogLik.vec(g_l0_t, AWX, which_known = NULL)
}
\arguments{
\item{g_l0_t}{vector c(gamma, lambda_0, theta)}

\item{AWX}{AWX dataframe}

\item{which_known}{if supplied, supposedly returns a partial likelihood where this is fixed.}
}
\value{
negative mean likelihood for the point g_l0_t
}
\description{
Wrapper for negLogLik to use with optim
}
