% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{nextArrivalCosine}
\alias{nextArrivalCosine}
\title{Next Arrival of inhomogeneous Poisson process}
\usage{
nextArrivalCosine(current_time, params)
}
\arguments{
\item{current_time}{the time on the clock.}

\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
time of the next arrival.
}
\description{
Generates  next arrival time (not inter-arrival!) of a nonhomogeneous Poisson process.
}
\examples{
params <- listParams(gamma=10,lambda_0=20,theta=2.5, eta = 1, mu = 1 , s = 3)
nextArrivalCosine(1.2,params = params)
}
