% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{oneKnownGrad}
\alias{oneKnownGrad}
\title{Title f}
\usage{
oneKnownGrad(which_known, params, AWX)
}
\arguments{
\item{which_known}{name of known parameter}

\item{params}{list of parameters}

\item{AWX}{dataset}
}
\value{
a gradient function for the two other parameter
}
\description{
Title f
}
