% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{oneKnownLik}
\alias{oneKnownLik}
\title{Function factory for partial likelihoods}
\usage{
oneKnownLik(which_known, params, AWX)
}
\arguments{
\item{which_known}{Which of the parameter values is \emph{known} to the estimator? "theta", "gamma" or "lambda_0".}

\item{params}{vector c(gamma, lambda_0, theta)}

\item{AWX}{data}
}
\value{
a function that computes likelihood for the remaining two parameter except which_known.
The function's input is always a vector where the parameters are ordered alphabetically by their names, gamma before lambda_0 before theta.
}
\description{
Function factory for partial likelihoods
}
\examples{
params <- exampleParams()
AWX <- resSimAWX(4,params)
f <- oneKnownLik("theta",params = params,AWX = AWX)
f(c(1,2)) # likelihood for gamma = 1 lambda_0 = 2
}
