% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pltArrivalsByTimeSegment}
\alias{pltArrivalsByTimeSegment}
\title{Plot the queue dynamics in a specified time interval}
\usage{
pltArrivalsByTimeSegment(segments = 24, RES, params)
}
\arguments{
\item{segments}{number of segments to divide one cycle into. Defaults to 24 (hours per day).}

\item{RES}{full simulations results list}

\item{params}{scenario parameters}
}
\description{
Plot the queue dynamics in a specified time interval
}
