% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pltQueueInterval}
\alias{pltQueueInterval}
\title{Plot the queue dynamics in a specified time interval}
\usage{
pltQueueInterval(from_to = c(1, 2), RES, params)
}
\arguments{
\item{from_to}{vector of start and stop of the time interval. Defaults to c(1,2).}

\item{RES}{full simulations results list}

\item{params}{scenario parameters}
}
\value{
plot only
}
\description{
Plot the queue dynamics in a specified time interval
}
\examples{
R <- exampleDataRES()
pltQueueInterval(RES = R, params = exampleParams())
}
