% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{pltRate}
\alias{pltRate}
\title{Plot the rate function}
\usage{
pltRate(params, n_cycles = 5)
}
\arguments{
\item{params}{names list of parameters (output of 'listParams()').}

\item{n_cycles}{of cycles to plot (defaults to 5)}
}
\description{
Plot the rate function
}
\examples{
pltRate(exampleParams())
}
