% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rateFactory}
\alias{rateFactory}
\title{Rate function factory}
\usage{
rateFactory(params)
}
\arguments{
\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
A function with argument t that computes the arrival rate.
}
\description{
Rate function factory
}
\examples{
params <- listParams(gamma=10,lambda_0=20,theta=2.5, eta = 1, mu = 1 , s = 3)
rate <- rateFactory(params)
rate(t = runif(3))
}
