% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RES2AWX}
\alias{RES2AWX}
\title{Utility: turn RES to AWX}
\usage{
RES2AWX(RES)
}
\arguments{
\item{RES}{results of resSimCosine()}
}
\value{
A data frame with columns A - inter-arrival times, W - waiting times, X - waiting time jump sizes
}
\description{
Utility: turn RES to AWX
}
