% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{resSimAWX}
\alias{resSimAWX}
\title{Simulation of big sample sizes, by parts}
\usage{
resSimAWX(n_thousands, params)
}
\arguments{
\item{n_thousands}{what is the desired n (in thousands)?}

\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
AWX dataframe
}
\description{
Simulation of big sample sizes, by parts
}
\examples{
params <-  exampleParams()
a <- Sys.time()
res <- resSimAWX(3, params = params)# 10 K
Sys.time() - a
}
