% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{resSimCosine}
\alias{resSimCosine}
\title{Atomic Simulation for periodic arrivals (cosine model)}
\usage{
resSimCosine(n, params)
}
\arguments{
\item{n}{Number of samples to generate.}

\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
A list with the queue data
}
\description{
Atomic Simulation for periodic arrivals (cosine model)
}
\examples{
params <- exampleParams()
R <- resSimCosine(n=100, params = params)
}
