% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{resSimCosineContinue}
\alias{resSimCosineContinue}
\title{Simulate results, possibly from given initial state}
\usage{
resSimCosineContinue(E0 = NULL, n, params)
}
\arguments{
\item{E0}{Optional argument - A list with results from a previous simulation}

\item{n}{Number of samples to generate.}

\item{params}{named list of parameters (output of 'listParams()').}
}
\value{
same as resSimCosine
}
\description{
Simulate results, possibly from given initial state
}
\examples{
 params <- exampleParams()
R <- resSimCosine(n=100, params = params)
R2 <- resSimCosineContinue(E0 = R, n = 100, params = params)
}
