% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scenarioParams}
\alias{scenarioParams}
\title{parameters by simulation scenario}
\usage{
scenarioParams(scenario_name = "C4")
}
\arguments{
\item{scenario_name}{"C1", "C2", "C3", "C4".}
}
\value{
params list with example number of servers
}
\description{
parameters by simulation scenario
}
\examples{
scenarioParams("C4")
}
