% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{twoKnownLik}
\alias{twoKnownLik}
\title{Function factory for partial likelihoods}
\usage{
twoKnownLik(which_not_known, params, AWX)
}
\arguments{
\item{which_not_known}{name of the UNKNOWN parameter to be estimated}

\item{params}{list of parameters}

\item{AWX}{AWX dataframe}
}
\value{
a function that computes likelihood for the the parameter which_not_known.
The function's input is always a vector where the parameters are ordered alphabetically by their names, gamma before lambda_0 before theta.
}
\description{
Function factory for partial likelihoods
}
