% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{VW}
\alias{VW}
\title{Liron's Virtual Waiting function}
\usage{
VW(Res.service, s)
}
\arguments{
\item{Res.service}{vector of residual service times}

\item{s}{the number of servers}
}
\description{
Liron's Virtual Waiting function
}
\examples{
VW(c(1.2,1.5,0.3), s = 2)
VW(c(1.2,1.5,0.3), s = 3)
VW(c(1.2,1.5,0.3), s = 4)
}
