% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{oneKnownGrad}
\alias{oneKnownGrad}
\title{Title f}
\usage{
oneKnownGrad(which_known, params, AWX)
}
\arguments{
\item{which_known}{Which of the parameter values is \emph{known} to the estimator? "theta", "gamma" or "lambda_0".}

\item{params}{vector c(gamma, lambda_0, theta)}

\item{AWX}{data}
}
\value{
a gradient function for the two other parameter
}
\description{
Title f
}
